Refactor health routes in to separate file
